requirement.txt
fastapi
uvicorn
requests
pydantic
python-dotenv
sentence-transformers
faiss-cpu
numpy
openai
